--- a/projects/miopen/addkernels/addkernels.cpp
+++ b/projects/miopen/addkernels/addkernels.cpp
@@ -33,6 +33,7 @@
 #include <sstream>
 #include <string>
 #include <vector>
+#include <cassert>
 
 namespace fs = miopen::fs;
 
@@ -92,6 +93,76 @@ void Bin2Hex(std::istream& source,
     }
 }
 
+void Bin2Asm(std::istream& source,
+             std::ostream& targetHeader,
+             std::ostream& targetAsm,
+             std::ostream& targetBin,
+             const fs::path& targetBinPath,
+             const std::string& variable,
+             bool nullTerminate,
+             size_t bufferSize)
+{
+    assert(!variable.empty());
+
+    source.seekg(0, std::ios::end);
+    const auto sourceSize = source.tellg();
+    source.seekg(0, std::ios::beg);
+
+    const auto buffer = std::make_unique<char[]>(bufferSize);
+
+    // Write header data
+    targetHeader << "extern \"C\" const size_t " << variable << "_SIZE;" << std::endl;
+    targetHeader << "extern \"C\" const char " << variable << "[" << sourceSize << "];" << std::endl;
+
+    const auto incbinOffset = targetBin.tellp();
+
+    // Write binary data
+    for(std::streamoff blockStart = 0; blockStart < sourceSize; blockStart += bufferSize)
+    {
+        source.read(buffer.get(), bufferSize);
+
+        const auto pos       = source.tellg();
+        const auto blockSize = (pos < 0 ? sourceSize : pos) - blockStart;
+
+        targetBin.write(buffer.get(), blockSize);
+    }
+
+    if(nullTerminate)
+        targetBin.put(0);
+
+    // Write assembly
+    const auto incbinSize = targetBin.tellp() - incbinOffset;
+    const auto writeSymbol = [&](const std::string& symbol, auto f)
+    {
+        targetAsm << "#if defined(_WIN32) || defined(__CYGWIN__)\n";
+        targetAsm << "     /* PE/COFF format */\n";
+        targetAsm << "     .section .rdata\n";
+        targetAsm << "#else\n";
+        targetAsm << "     /* ELF format */\n";
+        targetAsm << "     .section .note.GNU-stack,\"\",@progbits\n";
+        targetAsm << "     .type \"" << symbol << "\",@object\n";
+        targetAsm << "     .section .rodata\n";
+        targetAsm << "#endif\n";
+        targetAsm << "    .globl \"" << symbol << "\"\n";
+        targetAsm << "    .p2align 8\n";
+        targetAsm << symbol << ":\n";
+        f();
+        targetAsm << "#if !(defined(_WIN32) || defined(__CYGWIN__))\n";
+        targetAsm << "    .size " << symbol << ", . - " << symbol << "\n";
+        targetAsm << "#endif\n";
+    };
+
+    writeSymbol(variable, [&]
+    {
+        targetAsm << "    .incbin \"" << targetBinPath.string() << "\", " << incbinOffset << ", " << incbinSize << "\n";
+    });
+
+    writeSymbol(variable + "_SIZE", [&]
+    {
+        targetAsm << "    .quad " << incbinSize << "\n";
+    });
+}
+
 void PrintHelp()
 {
     std::cout << "Usage: addkernels {<option>}" << std::endl;
@@ -110,6 +181,8 @@ void PrintHelp()
     std::cout << "           -m[ark-includes] : mark variables that represent include files with "
                  "'_INCLUDE'. Default: off"
               << std::endl;
+    std::cout << "           -a[sm] <asm> <bin> : inline source via assembly file and binary "
+              << "instead of header. Default: off" << std::endl;
 }
 
 [[noreturn]] void WrongUsage(std::string_view error)
@@ -133,7 +206,10 @@ void Process(const fs::path& sourcePath,
              size_t lineSize,
              bool recurse,
              bool as_extern,
-             bool mark_includes)
+             bool mark_includes,
+             std::ofstream& asmStream,
+             std::ofstream& binStream,
+             const fs::path& targetBinPath)
 {
     if(!fs::exists(sourcePath))
     {
@@ -200,7 +276,12 @@ void Process(const fs::path& sourcePath,
         variable = "MIOPEN_KERNEL_" + variable;
     }
 
-    Bin2Hex(*source, target, variable, true, bufferSize, lineSize);
+    if (asmStream.is_open()) {
+        assert(binStream.is_open());
+        Bin2Asm(*source, target, asmStream, binStream, targetBinPath, variable, true, bufferSize);
+    } else {
+        Bin2Hex(*source, target, variable, true, bufferSize, lineSize);
+    }
 }
 
 int main(int argc, char* argv[])
@@ -219,6 +300,8 @@ int main(int argc, char* argv[])
     size_t lineSize   = 16;
 
     std::string targetFile;
+    std::string asmOutputFile;
+    std::string binOutputFile;
     std::vector<fs::path> sourceFiles;
 
     bool recurse       = true;
@@ -270,6 +353,25 @@ int main(int argc, char* argv[])
         {
             as_extern = true;
         }
+        else if (arg == "-a" || arg == "-asm")
+        {
+            if (i + 2 >= argc)
+            {
+                std::ostringstream ss;
+                ss << arg << " requires arguments <asm path> <bin path>";
+                WrongUsage(ss.str());
+            }
+
+            std::string outputAsm{argv[++i]};
+            if (!asmOutputFile.empty())
+                std::cerr << "Warning: overriding asm output file\n    '" << asmOutputFile
+                          << "'\nwith\n    '" << outputAsm << "'\n";
+
+            asmOutputFile = outputAsm;
+
+            std::string outputBin{argv[++i]};
+            binOutputFile = outputBin;
+        }
         else
         {
             UnknownArgument(arg);
@@ -290,9 +392,39 @@ int main(int argc, char* argv[])
     ss << "#ifndef MIOPEN_USE_CLANG_TIDY\n"
           "#include <cstddef>\n";
 
+    std::ofstream asmStream;
+    std::ofstream binStream;
+    if (!asmOutputFile.empty())
+    {
+        assert(!binOutputFile.empty());
+        asmStream.open(asmOutputFile, std::ios::out | std::ios::binary);
+        binStream.open(binOutputFile, std::ios::out | std::ios::binary);
+
+        if (!asmStream.is_open())
+        {
+            std::cerr << "failure opening file: " << asmOutputFile << "\n";
+            return 1;
+        }
+
+        if (!binStream.is_open())
+        {
+            std::cerr << "failure opening file: " << binOutputFile << "\n";
+            return 1;
+        }
+    }
+
     for(const auto& file : sourceFiles)
     {
-        Process(file, ss, bufferSize, lineSize, recurse, as_extern, mark_includes);
+        Process(file,
+                ss,
+                bufferSize,
+                lineSize,
+                recurse,
+                as_extern,
+                mark_includes,
+                asmStream,
+                binStream,
+                binOutputFile);
     }
 
     ss << "#endif\n";