## Motivation

Before these changes, it took a really long time to recompile MIOpen after changing the kernels. This breaks my zoomer brain, as I get too distracted while waiting for the 15 second compile times. I've improved it down to only a few seconds.

Some statistics:

before, clean, excluding reconfigure: 162.30 sec
after fixing CMakeLists.txt dependency, clean, excluding reconfigure: 162.30 sec
after previous + incbin, clean, excluding configure: 133.67 secs
after previous + using mold, clean, including reconfigure: 130.79 secs

before, incremental, including reconfigure: 16.44 secs (I swear this took 60+ seconds in earlier testing)
after fixing CMakeLists.txt dependency, incremental, now excluding reconfigure: 11.80 secs
after previous + incbin, incremental, excluding reconfigure: 5.81 secs
after previous + using mold, incremental, excluding reconfigure: 2.33 secs

By Mold i mean the [Mold linker](https://github.com/rui314/mold). It can be used with `-DCMAKE_LINKER=MOLD`. This gives an error if done during a clean build, the trick is to reconfigure an already configured build directory using this flag. This is an entirely optional step, so I didn't look into it too much.

## Technical Details

There are 2 main problems causing long rebuild times:
* An unneeded dependency from all MIOpen kernels onto the CMakeLists.txt. This is was there due to historical reasons; the CMake script used to be also used to embed kernels into the the library. The `addkernels` utility is used for that now, so this is no longer needed. See commit https://github.com/ROCm/rocm-libraries/commit/9675cb1a81827b1f79b4f4b94277e9707605331f for more details.
* Kernels are embedded by converting them to array literals and compiling. This is pretty slow, especially with GCC, as we've also found out with rocRAND. With rocRAND, we've added an optional `incbin`-based strategy for embedding kernels, which is much faster. To this end, I've added a new option to `addkernels` to generate assembly files and binaries instead. The assembly file is based on the same system that we use in rocRAND, see [here](https://github.com/ROCm/rocm-libraries/commit/3fd2982a785bb7c5530504049713c23f5c35d0a6). While this option should be cross-platform, I'm not 100% sure about the `dllexport` options, and so I've hidden the use of `incbin`-based including behind the `-DMIOPEN_INCBIN` flag for now.

## Test Plan

Existing tests should cover the 2nd case, assuming testing is done with `-DMIOPEN_INCBIN=ON` enabled. The 1st case was tested locally, I don't really see a good way to test that anyway.

## Test Result

`HSA_ENABLE_SDMA=0 ctest --output-on-failure -R "^test_" --parallel 16` passes on my machine after setting `-DMIOPEN_INCBIN=ON`. I don't know about Windows.